// src/app/models/profile.model.ts
export class Profile {
    accountId: number;
    personaName: string;
    name: string;
    plus: boolean;
    cheese: number;
    steamId: number;
    avatar: string;
    avatarMedium: string;
    avatarFull: string;
    profileUrl: string;
    lastLogin: Date;
    locCountryCode: number;
    isContributor: boolean;
}